Add http_cont

+++ src/aos_string.h
@@ -13,7 +13,7 @@ typedef struct {
 #define aos_string(str)     { sizeof(str) - 1, (char *) str }
 #define aos_null_string     { 0, NULL }
 #define aos_str_set(str, text)                                  \
-    (str)->len = strlen(text); (str)->data = (char *) text
+    (str)->len = (int)strlen(text); (str)->data = (char *) text
 #define aos_str_null(str)   (str)->len = 0; (str)->data = NULL
 
 #define aos_tolower(c)      (char) ((c >= 'A' && c <= 'Z') ? (c | 0x20) : c)
